% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nhoods.R
\name{calcNhoodMean}
\alias{calcNhoodMean}
\title{Average expression values within each neighbourhood}
\usage{
calcNhoodMean(X, mask)
}
\arguments{
\item{X}{Matrix of expression values}

\item{mask}{Indicator matrix specifying cell groupings}
}
\description{
Average expression values within each neighbourhood
}
