% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{downsampleSCE}
\alias{downsampleSCE}
\title{Uniformly downsample a SingleCellExperiment object}
\usage{
downsampleSCE(sce, ncells)
}
\arguments{
\item{sce}{SingleCellExperiment object to downsample}

\item{ncells}{Number of cells required}
}
\value{
Downsampled SingleCellExperiment object
}
\description{
Uniformly downsample a SingleCellExperiment object
}
