% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{exportToGMT}
\alias{exportToGMT}
\title{Write colData to a GMT file}
\usage{
exportToGMT(sce, group, gmt_path)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{group}{colData observation to export}

\item{gmt_path}{Path to write GMT file}
}
\description{
Used to export cell sets for Waddington-OT analysis.
}
\details{
The gmt format consists of one cell set per line. Each line is a
tab-separated list composed as follows : \cr
\itemize{
\item The set name (can contain spaces) \cr
\item A commentary / description of the set (may be empty or contain spaces) \cr
\item A tab-separated list of set members
}
}
