% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{filterSharedFeatures}
\alias{filterSharedFeatures}
\title{Obtain one-to-one orthologs from Ensembl.}
\usage{
filterSharedFeatures(
  query_genes,
  query_species,
  one2one.data,
  outputIDs = TRUE
)
}
\arguments{
\item{query_genes}{Genes from which to find one-to-one orthologs.}

\item{query_species}{Species to find one-to-one orthologs with provided in
Ensembl format (e.g. "hsapiens", "mmusculus").}

\item{one2one.data}{A data.frame of results retrieved from the Ensembl BioMart
database containing "_homolog_ensembl_gene" or  "_homolog_associated_gene_name"
fields.}

\item{outputIDs}{Specifies whether to return the one-to-one orthologs in
Ensembl gene ID format or using their gene names.}
}
\description{
Obtain one-to-one orthologs from Ensembl.
}
