% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{getEnsemblHomologs}
\alias{getEnsemblHomologs}
\title{Queries Ensembl BioMart database for homologs between species.}
\usage{
getEnsemblHomologs(ref_species, ref_genes, query_species)
}
\description{
Queries Ensembl BioMart database for homologs between species.
}
