% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGenesFromGOTerms}
\alias{getGenesFromGOTerms}
\title{Used to obtain genes related to apoptosis and proliferation for Waddington-OT analysis
Positively regulates...
  cell_death <- "GO:0010942"
  cell_population_proliferation <- "GO:0008284"}
\usage{
getGenesFromGOTerms(species, go_terms, return_value = "ensembl_gene_id")
}
\description{
Used to obtain genes related to apoptosis and proliferation for Waddington-OT analysis
Positively regulates...
  cell_death <- "GO:0010942"
  cell_population_proliferation <- "GO:0008284"
}
