% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSharedUniqueMarkers}
\alias{getSharedUniqueMarkers}
\title{Get dataframe of shared and unique markers.
Used for mouse/rabbit cell type markers heatmap.}
\usage{
getSharedUniqueMarkers(merge.markers, N = 1)
}
\description{
Get dataframe of shared and unique markers.
Used for mouse/rabbit cell type markers heatmap.
}
