% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{isEnsemblID}
\alias{isEnsemblID}
\title{Checks if provided genes are in Ensembl gene ID format.
 @export}
\usage{
isEnsemblID(genes)
}
\description{
Checks if provided genes are in Ensembl gene ID format.
 @export
}
