% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_annotation.R
\name{loadReadsAroundGRange}
\alias{loadReadsAroundGRange}
\title{Obtains reads within a window of a given GRange}
\usage{
loadReadsAroundGRange(bam_file, gene_range, offset = 1e+06)
}
\arguments{
\item{bam_file}{Rsamtools BamFile object}

\item{gene_range}{GRange object of gene of interest}

\item{offset}{Offset from the start and end of the specified gene in which to
consider reads}
}
\description{
Constrains the reads considered to within a specified window of a GRange.
}
