% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/plot_nhoods.R
\name{plotNhoodSimGroups}
\alias{plotNhoodSimGroups}
\title{sim_values in order of milo nhoods
TODO: Make into separate functions so can pass plot specific params using ...}
\usage{
plotNhoodSimGroups(
  milo,
  sim_values,
  group_by = "celltype",
  facet_by = NULL,
  type = "ridge",
  orientation = "vertical",
  colour_by = NULL,
  size = 0.2,
  subset = NULL,
  decreasing = FALSE,
  rel_min_height = 0,
  show_rank = FALSE,
  rank_size = 2,
  group_colours = scrabbitr::getCelltypeColours(),
  xlabel = "Cell type",
  ylabel = "Correlation"
)

plotNhoodSimGroups(
  milo,
  sim_values,
  group_by = "celltype",
  facet_by = NULL,
  type = "ridge",
  orientation = "vertical",
  colour_by = NULL,
  size = 0.2,
  subset = NULL,
  decreasing = FALSE,
  rel_min_height = 0,
  show_rank = FALSE,
  rank_size = 2,
  group_colours = scrabbitr::getCelltypeColours(),
  xlabel = "Cell type",
  ylabel = "Correlation"
)
}
\description{
sim_values in order of milo nhoods
TODO: Make into separate functions so can pass plot specific params using ...

sim_values needs to be in order of milo nhoods
TODO: Split into smaller functions
}
