% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_annotation.R
\name{plotReadDistsNearGene}
\alias{plotReadDistsNearGene}
\title{Plots histogram of distances of reads that are nearest to the gene given}
\usage{
plotReadDistsNearGene(gene_id, bam_file, gtf_gf)
}
\description{
Plots histogram of distances of reads that are nearest to the gene given
}
