% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{prepareEdgeR}
\alias{prepareEdgeR}
\title{Prepare groupings in SingleCellExperiment object for running edgeR}
\usage{
prepareEdgeR(sce, group_by, group1, group2, block_by)
}
\description{
Prepare groupings in SingleCellExperiment object for running edgeR
}
