% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp.R
\name{runEdgeR}
\alias{runEdgeR}
\title{Compute differentially expressing genes using EdgeR}
\usage{
runEdgeR(
  sce,
  min_detection_rate_per_group = 0.1,
  min.logFC = 2,
  threshold_fdr = 0.1
)
}
\description{
Compute differentially expressing genes using EdgeR
}
